#3 Renamed con.rs to connection.rs
